+++ src/types/apiType.ts
@@ -1,0 +1,10 @@
+export interface Login {
+  id: number;
+  username: string;
+  password: string;
+  token: string;
+}
+
+export interface ApiConfig {
+  baseUrl: string;
+}
